Refactor request approval components and consolidate functionality
- Introduce ApproveTable component to handle request approvals, replacing EditorContentTableApprove and RaContentTableApprove.
- Update RequestApprove component to conditionally render ApproveTable based on user role.
- Enhance data fetching and state management for request approvals using Redux.
- Improve UI layout and responsiveness for the new approval table.
- Remove deprecated components to streamline the codebase.
